feat(silence): Allowing silence feature in all groups, disabled by default

+++ src/chat/group/screens/screen_silence_end.py
@@ -18,10 +18,16 @@ async def manage(
     :return: None
     """
 
+    if not group_chat.is_muted:
+        await full_message_send(
+            context, phrases.SILENCE_NOT_ACTIVE, update=update, add_delete_button=True
+        )
+        return
+
     group_chat.is_muted = False
     group_chat.save()
 
     UnmutedUser.delete().where(UnmutedUser.group_chat == group_chat).execute()
 
     # Confirmation message
-    await full_message_send(context, phrases.SILENCE_END, update=update)
+    await full_message_send(context, phrases.SILENCE_END, update=update, add_delete_button=True)
